# docs/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "CodeSage API",
    "description": "Local-first, stateless AI pull request review assistant.\n\n## Authentication\nNo user authentication. Configure `GitHub:PersonalAccessToken` and `OpenAI:ApiKey`\nvia user-secrets or environment variables.\n\n## Errors\nAll failures return RFC 7807 `application/problem+json` with an `errorCode` extension.\n\n## Versioning\nCurrent surface is `/api/v1`. Breaking changes will ship as `/api/v2`.",
    "contact": {
      "name": "CodeSage Contributors",
      "url": "https://github.com/codesage"
    },
    "license": {
      "name": "MIT"
    },
    "version": "v1"
  },
  "paths": {
    "/api/v1/configuration": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "Public configuration summary",
        "description": "Returns non-secret runtime configuration useful for clients and troubleshooting. API keys and tokens are never included — only whether they are configured.",
        "operationId": "GetConfigurationV1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationSummaryResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — RFC 7807 ProblemDetails with an errors object.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/system/status": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "System diagnostics",
        "description": "Returns configuration readiness and optional connectivity probes. Never includes API keys or tokens. Disable probes with Application:ProbeExternalConnectivity=false.",
        "operationId": "GetSystemStatusV1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — RFC 7807 ProblemDetails with an errors object.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigurationSummaryResponse": {
        "type": "object",
        "properties": {
          "application": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "gitHubApiBaseUrl": {
            "type": "string"
          },
          "gitHubTokenConfigured": {
            "type": "boolean"
          },
          "aiProvider": {
            "type": "string"
          },
          "aiModel": {
            "type": "string"
          },
          "openAiBaseUrl": {
            "type": "string"
          },
          "openAiApiKeyConfigured": {
            "type": "boolean"
          },
          "probeExternalConnectivity": {
            "type": "boolean"
          },
          "requireSecretsAtStartup": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "description": "Public configuration summary for clients. Never includes secrets.",
        "example": "{\"application\":\"CodeSage\",\"version\":\"0.1.0\",\"environment\":\"Development\",\"gitHubApiBaseUrl\":\"https://api.github.com/\",\"gitHubTokenConfigured\":true,\"aiProvider\":\"OpenAI\",\"aiModel\":\"gpt-4o-mini\",\"openAiBaseUrl\":\"https://api.openai.com/v1/\",\"openAiApiKeyConfigured\":true,\"probeExternalConnectivity\":true,\"requireSecretsAtStartup\":true}"
      },
      "ConnectivityCheckResult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "httpStatusCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Result of an optional external connectivity probe."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "SystemStatusResponse": {
        "type": "object",
        "properties": {
          "application": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "aiProvider": {
            "type": "string"
          },
          "aiModel": {
            "type": "string"
          },
          "gitHubTokenConfigured": {
            "type": "boolean"
          },
          "openAiApiKeyConfigured": {
            "type": "boolean"
          },
          "gitHubConnectivity": {
            "$ref": "#/components/schemas/ConnectivityCheckResult"
          },
          "openAiConnectivity": {
            "$ref": "#/components/schemas/ConnectivityCheckResult"
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false,
        "description": "Local diagnostics for troubleshooting configuration (never includes secrets).",
        "example": "{\"application\":\"CodeSage\",\"version\":\"0.1.0\",\"environment\":\"Development\",\"aiProvider\":\"OpenAI\",\"aiModel\":\"gpt-4o-mini\",\"gitHubTokenConfigured\":true,\"openAiApiKeyConfigured\":true,\"gitHubConnectivity\":{\"status\":\"Skipped\",\"message\":\"External probes disabled\"},\"openAiConnectivity\":{\"status\":\"Skipped\",\"message\":\"External probes disabled\"},\"diagnostics\":[\"Configuration looks ready for local reviews.\"]}"
      }
    }
  },
...
}